# 8317678: Fix up hashCode() for ZipFile.Source.Key