… test/micro/org/openjdk/bench/java/util/zip/ZipFileOpen.java