Fix up java.util.zip.ZipFile$Source hashCode() impl so that duplicate Source objects aren't created for the same zip file.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8317678](https://bugs.openjdk.org/browse/JDK-8317678): Fix up hashCode() for ZipFile.Source.Key (**Bug** - P3)


### Reviewers
 * [Lance Andersen](https://openjdk.org/census#lancea) (@LanceAndersen - **Reviewer**) ⚠️ Review applies to [ff6c7d70](https://git.openjdk.org/jdk/pull/16115/files/ff6c7d7050125f327fdbfba4431e74471886db9c)
 * [Alan Bateman](https://openjdk.org/census#alanb) (@AlanBateman - **Reviewer**) ⚠️ Review applies to [3635e542](https://git.openjdk.org/jdk/pull/16115/files/3635e5424c5c385396896c5b0d4a0b49abf235f1)
 * [Jaikiran Pai](https://openjdk.org/census#jpai) (@jaikiran - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/16115/head:pull/16115` \
`$ git checkout pull/16115`

Update a local copy of the PR: \
`$ git checkout pull/16115` \
`$ git pull https://git.openjdk.org/jdk.git pull/16115/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 16115`

View PR using the GUI difftool: \
`$ git pr show -t 16115`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/16115.diff">https://git.openjdk.org/jdk/pull/16115.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/16115#issuecomment-1754706555)