@AlanBateman 2023-10-10
Can you look at the equals method too, need to decide if the checking of lastModifiedTime should be removed.