@coffeys 2023-10-10
Thanks for the comments to date. I think the latest commit should satisfy requests made.

for `Key.equals(Object)` and `Key.hashCode()` - I've changed the impl so that attrs.lastModifiedTime() is only used in the case where `attrs.fileKey()` returns null. Is that in the direction you were suggesting Alan ?