@coffeys 2023-10-13
`lastModifiedTime()` was used in the native code and brought across when native code was removed in JDK-8145260

I thought we might be able to drop it for the case where we have a unique file key but that won't cover the case where an existing zip file is opened and modified. Re-introduced the lastModifiedTime checks as a result. We need to be able to detect such scenarios, otherwise new ZipFile constructs would continue to use a stale <Key, Source> mapping (if such Keys still existed)

I've updated the test case to junit format and beefed it up with some extra scenarios to cover the above.