@coffeys 2023-10-17
Removed the `getCanonicalFile `calls. The patch now optimizes the case where fileKey() is supported. Updated the ZipFileOpen.java benchmark test case also. before and after benchmark results for linux and windows posted below

```
linux:
without patch:

Benchmark                       (size)  Mode  Cnt       Score       Error  Units
ZipFileOpen.openCloseZipFile       512  avgt   15   75378.890 ?   642.370  ns/op
ZipFileOpen.openCloseZipFile      1024  avgt   15  150814.293 ? 29801.424  ns/op
ZipFileOpen.openCloseZipFilex2     512  avgt   15  168104.286 ? 41505.778  ns/op
ZipFileOpen.openCloseZipFilex2    1024  avgt   15  299282.488 ? 44496.777  ns/op
Finished running test 'micro:java.util.zip.ZipFileOpen'

with patch
Benchmark                       (size)  Mode  Cnt       Score       Error  Units
ZipFileOpen.openCloseZipFile       512  avgt   15   82715.346 ? 21700.950  ns/op
ZipFileOpen.openCloseZipFile      1024  avgt   15  142253.235 ?  1054.553  ns/op
ZipFileOpen.openCloseZipFilex2     512  avgt   15   77238.320 ?   194.585  ns/op
ZipFileOpen.openCloseZipFilex2    1024  avgt   15  161828.411 ? 45486.890  ns/op
Finished running test 'micro:java.util.zip.ZipFileOpen'

windows:
without patch:
ZipFileOpen.openCloseZipFile       512  avgt   15  319543.232 ±  46112.481  ns/op
ZipFileOpen.openCloseZipFile      1024  avgt   15  463719.381 ± 139409.316  ns/op
ZipFileOpen.openCloseZipFilex2     512  avgt   15  750838.029 ± 190726.572  ns/op
ZipFileOpen.openCloseZipFilex2    1024  avgt   15  875174.035 ± 140971.036  ns/op

with patch:
Benchmark                       (size)  Mode  Cnt       Score        Error  Units
ZipFileOpen.openCloseZipFile       512  avgt   15  400393.718 ± 161890.652  ns/op
ZipFileOpen.openCloseZipFile      1024  avgt   15  380922.933 ±  21776.006  ns/op
ZipFileOpen.openCloseZipFilex2     512  avgt   15  623589.801 ±  38110.510  ns/op
ZipFileOpen.openCloseZipFilex2    1024  avgt   15  774654.146 ±  50413.718  ns/op
```